% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_helpers.R
\name{acs_clac}
\alias{acs_clac}
\alias{acs_sum}
\alias{acs_se_sum}
\alias{acs_se_prop}
\alias{acs_se_ratio}
\alias{acs_se_prod}
\title{Calculate Derived Variables and Standard Errors for ACS Estimates}
\usage{
acs_sum(..., na.rm = TRUE, .envir = parent.frame())

acs_se_sum(..., na.rm = TRUE, .envir = parent.frame())

acs_se_prop(num, num_se, denom, denom_se)

acs_se_ratio(num, num_se, denom, denom_se)

acs_se_prod(val_1, se_1, val_2, se_2)
}
\arguments{
\item{...}{[\code{expressions}]: Expressions string(s) to format into ACS
variable codes with \code{\link[=acs_vars]{acs_vars()}} or strings corresponding to column names.
are concatenated together before formatting.}

\item{na.rm}{[\code{logical: TRUE}]: A logical value indicating whether NA values
should be stripped before the computation proceeds.}

\item{.envir}{[\code{environment: parent.frame()}]: Environment to evaluate each
expression in. Expressions are evaluated from left to right. If .x is an
environment, the expressions are evaluated in that environment and .envir
is ignored.}

\item{num}{Estimate for the numerator.}

\item{num_se}{Standard error for the numerator.}

\item{denom}{Estimate for the denominator.}

\item{denom_se}{Standard error for the denominator.}

\item{val_1}{Estimate for the first value of product}

\item{se_1}{Stanadard error for the first value of product}

\item{val_2}{Estimate for the first value of product}

\item{se_2}{Stanadard error for the first value of product}
}
\description{
A collection of functions to help calculate derived variables and standard
errors from the ACS estimates returned by \code{\link[=acs_transform]{acs_transform()}}.
}
\details{
\itemize{
\item \code{acs_sum()}: Sum rows within \code{\link[dplyr:mutate]{dplyr::mutate()}} using \code{\link[=acs_vars]{acs_vars()}} syntax to list ACS columns.
\item \code{acs_se_sum()}: Calculate the standard error for the sum (or difference) of estimates
\item \code{acs_se_prop()}: Calculate the standard error for a proportion
\item \code{acs_se_ratio()}: Calculate the standard error for a ratio or mean (that isn't a proportion)
\item \code{acs_se_prod()}: Calculate the standard error for the product of two estimates
}
}
\examples{

library(dplyr)

df <- tibble(
  b25070_e001 = c(1, 2, 3),
  b25070_e002 = c(3, NA, 1),
  b25070_e003 = c(1, 2, NA)
)

df \%>\%
  mutate(
    foo = acs_sum("b25070_e{1:2*}"),
    bar = acs_sum("b25070_e{1:2*}", "b25070_e003"),
    baz = acs_sum("b25070_e002", "b25070_e003")
  )

}
\seealso{
\code{\link[=acs_vars]{acs_vars()}}, \code{\link[=acs_margin_to_se]{acs_margin_to_se()}}
}
