% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tables.R
\name{acs_create_indexes}
\alias{acs_create_indexes}
\title{Create Index on ACS Table in Database}
\usage{
acs_create_indexes(conn, table_name)
}
\arguments{
\item{conn}{\[`DBIConnection`]:A DBI connection object obtained from
`DBI::dbConnect()`.}

\item{table_name}{\[`character`]:The name of the table in the database that
has been created by `acssf::acs_make_table()`.}
}
\description{
This function creates a unique index (if it does not exist) on a ACS table
in a database created by `acssf::acs_make_table()`.
}
