% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_helpers.R
\name{acs_drop}
\alias{acs_drop}
\alias{acs_drop_estimates}
\alias{acs_drop_margins}
\title{Drop ACS variable columns from dataframe}
\usage{
acs_drop_estimates(df)

acs_drop_margins(df)
}
\arguments{
\item{df}{\[`data.frame`]: A data.frame with ACS variable columns using the
format created by [acs_make_table()]: `b25002_e001` for estimates and
`b25002_m001` for margins.}
}
\description{
Simple wrapper of [dplyr::select()] to drop ACS variable columns (estimates
or margins) created by [acssf::make_raw_table()].
}
\details{
This function simply wraps the `dplyr` expression `select(df,
  -maches("pattern"))`, where the regular expression is
  `"[bc]\\d{5}[a-z]*_e\\d{3}"` for estiamtes and `"[bc]\\d{5}[a-z]*_m\\d{3}"`
  for margins.
}
\examples{

library(dplyr)
library(tibble)

df <- tibble(
  endyear = 2009L,
  span = 1L,
  b25070_e001 = c(1, 2, 3),
  b25070_m001 = c(0.2, 0.5, 0.7),
  b25070_e002 = c(3, NA, 1),
  b25070_m002 = c(0.1, NA, 0.4)
)

acs_drop_estimates(df)

acs_drop_margins(df)

df \%>\%
  acs_drop_margins() \%>\%
  mutate(foo = acs_sum("b25070_e{1:2*}")) \%>\%
  acs_drop_estimates()

}
