% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_extract_raw.R
\name{acs_extract_raw}
\alias{acs_extract_raw}
\title{Create csv file/dataframe of ACS variables for a given sample and geography}
\usage{
acs_extract_raw(acs_dir, endyear, span, geo, sum_level, vars_table)
}
\arguments{
\item{acs_dir}{\[`character(1)`]: The root directory in which all the ACS
data has been downloaded with [acs_download()].}

\item{endyear}{\[integer(1)]: The endyear of the desired ACS sample. For
example, use 2010 for the 2010 1-year ACS or the 2006-2010 5-yer ACS.}

\item{span}{\[integer(1)]: The span of years for ACS estimates. ACS contains
1-, 3-, and 5-year surveys.}

\item{geo}{\[`charater(1)`]: The 2-letter state abbreviation or 2-digit FIPS
code for the state for which data will be downloaded. For geogrpahies that
do not nest within states, use `"us"`.}

\item{sum_level}{\[`character(1)`]: The Census Bureau code for the summary
level to include in the table. (eg. "010" = United States) For full list
see <https://factfinder.census.gov/help/en/summary_level_code_list.htm>.}

\item{vars_table}{\[`data.frame`]: A dataframe with a single column
containing the desired ACS variable codes to be included in the output
table. The ACS codes must be stored as a character column using the format
`"b25003_001"`.}
}
\description{
This function parses the data downloded by [acs_download()] and creates a csv
file for the selected parameters.
}
