% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_make_indicators.R
\name{acs_make_indicators}
\alias{acs_make_indicators}
\title{Create indicators from clean ACS SF data}
\usage{
acs_make_indicators(acs_dir, out_dir = NULL, endyear, span, sum_level,
  .f = NULL)
}
\arguments{
\item{acs_dir}{\[`character(1)`]: The root directory in which all the ACS SF data will be saved.}

\item{out_dir}{\[`character(1)`]: The directory to save indicator files to. Defults to "`acs_dir`/Output".}

\item{endyear}{\[`integer(1)`]: The endyear of the ACS sample. 2005 through 2016 are
available.}

\item{span}{\[`integer(1)`]: The span of years for ACS estimates. ACS 1-year, and
5-year surveys are supported.}

\item{sum_level}{\[`character(1)`]: The Census summary level code. For full list of supported summary levels see `sum_level_info`.}

\item{.f}{A function or formula to be passed to [`purrr::as_mapper()`]. The function must always return a dataframe.

  If a __function__, it is used as is.

  If a __formula__, e.g. `~ .x + 2`, it is converted to a function. There
  are three ways to refer to the arguments:

  * For a single argument function, use `.`
  * For a two argument function, use `.x` and `.y`
  * For more arguments, use `..1`, `..2`, `..3` etc

  This syntax allows you to create very compact anonymous functions.}
}
\description{
Import all csv files for a given geography type, endyear, and span that were created by [acs_extract_raw()], and apply a function to each file to further process data from.
}
