% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tables.R
\name{acs_make_table}
\alias{acs_make_table}
\title{Create ACS Table in Database}
\usage{
acs_make_table(acs_dir, endyear, span, geo, sumlevels, table_name, conn)
}
\arguments{
\item{acs_dir}{The root directory in which all the ACS SF data will be saved.}

\item{endyear}{The endyear of the ACS sample. 2005 through 2016 are
available.}

\item{span}{The span of years for ACS estimates. ACS contains 1-, 3-, and
5-year surveys.}

\item{geo}{The 2-letter abbreviation for the state for which data will be
downloaded. For geogrpahies that do not nest within states, use `"us"`.}

\item{sumlevels}{[`character`]: A character vector of the Census Bureau summary levels to include in the table. (eg. "010" = United States) For full list see https://factfinder.census.gov/help/en/summary_level_code_list.htm}

\item{table_name}{\[`character`]:The name of the table in the database that should be create or appended to.}

\item{conn}{\[`DBIConnection`]:A DBI connection object obtained from `DBI::dbConnect()`.}
}
\description{
This function parses the data downloded by `acssf::download_acs()` and either creates
or appends to a table in a database with American Community Survey (ACS)
Summary File (SF) data.
}
