% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_helpers.R
\name{acs_margin_to_se}
\alias{acs_margin_to_se}
\title{Convert All Margin of Error columns to Standard Errors}
\usage{
acs_margin_to_se(.data, level = 0.9)
}
\arguments{
\item{.data}{Dataframe as created by \code{\link[=acs_transform]{acs_transform()}} before any changes
have been made.}

\item{level}{Confidence level for standard errors. Defaults to 0.90.}
}
\description{
By default ACS data has columns for estimates and their associated margins of
error (at the 90% level). When calculating new variables from ACS estiamtes
it's often easier to work with standard errors rather than margins of error.
This function will convert every margin of error column (eg. \code{b25064_m001})
into standard errors for a given confidence level (eg. \code{b25064_se001}). Once
this conversion has been made there are a variety of \link[=acs_se_sum]{standard error calculation helpers} available in this package.
}
