% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_sum.R
\name{acs_moe_sum}
\alias{acs_moe_sum}
\title{Calculate the margin of error for a derived sum of ACS variables}
\usage{
acs_moe_sum(..., na.rm = TRUE, .envir = parent.frame())
}
\arguments{
\item{na.rm}{[\code{logical: TRUE}]: A logical value indicating whether NA values
should be stripped before the computation proceeds.}

\item{.envir}{[\code{environment: parent.frame()}]: Environment to evaluate each
expression in. Expressions are evaluated from left to right. If .x is an
environment, the expressions are evaluated in that environment and .envir
is ignored.}
}
\description{
This function calculates the margin of error for a derived sum of ACS
variables from the ACS estimates and margins of error returned by
\code{\link[=acs_transform]{acs_transform()}}. It assumes the data structure of a dataframe returned by
\code{\link[=acs_transform]{acs_transform()}}, where each row is a geography and each column is an ACS
estimate or margin of error, and so is used within a \code{\link[dplyr:mutate]{dplyr::mutate()}} call
and sums across rows. The main argument to this function is first parsed with
\code{\link[=acs_vars]{acs_vars()}} and then the columns are found within the dataframe. The
variables provided should be fore the estimates and it will also find the
corresponding margins of error columns using the standard naming conversion
of \code{\link[=acs_transform]{acs_transform()}}.
}
