% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_helpers.R
\name{acs_vars}
\alias{acs_vars}
\title{Create ACS variable codes}
\usage{
acs_vars(..., .envir = parent.frame())
}
\arguments{
\item{...}{[\code{expressions}]: Expressions string(s) to format, multiple inputs
are concatenated together before formatting.}

\item{.envir}{[\code{environment: parent.frame()}]: Environment to evaluate each
expression in. Expressions are evaluated from left to right. If .x is an
environment, the expressions are evaluated in that environment and .envir
is ignored.}
}
\description{
Use \code{\link[glue:glue]{glue::glue()}} syntax to format and interpolate strings into ACS variable
codes. By placing * at the end of a glue expression the values will be padded
with \code{"0"} to length 3.
}
\examples{
acs_vars("b25003_e{1:3*}")

acs_vars("b25003_e{c(1, 3)*}")

}
\seealso{
\url{http://glue.tidyverse.org/articles/transformers.html} for
explanation of transformers used here to pad numbers.
}
