% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acssf-package.R
\docType{package}
\name{acssf-package}
\alias{acssf}
\alias{acssf-package}
\title{Download ACS Summary File Data and Build Database.}
\description{
Download ACS Summary File data from FTP site, clean data and create tables in
database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/austensen/acssf}
  \item Report bugs at \url{https://github.com/austensen/acssf/issues}
}

}
\author{
\strong{Maintainer}: Maxwell Austensen \email{fmaxwell.austensen@nyu.edu}

}
