% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_acs.R
\name{download_acs}
\alias{download_acs}
\title{Download ACS Summary File Data}
\usage{
download_acs(acs_dir, endyear, span, geo, overwrite = FALSE)
}
\arguments{
\item{acs_dir}{The root directory in which all the ACS SF data will be saved.}

\item{endyear}{The endyear of the ACS sample. 2005 through 2016 are
available.}

\item{span}{The span of years for ACS estimates. ACS contains 1-, 3-, and
5-year surveys.}

\item{geo}{The 2-letter abbreviation for the state for which data will be
downloaded. For geogrpahies that do not nest within states, use `"us"`.}

\item{overwrite}{Whether existing versions of these files be overwriten.
Defaults to `FALSE`.}
}
\description{
This function downloads American Community Survey (ACS) Summary File (SF)
data from the Census Bureau's FTP site.
}
