% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_download.R
\name{download_docs}
\alias{download_docs}
\title{Download Seq/Table/Var Info}
\usage{
download_docs(docs_dir, year, span)
}
\arguments{
\item{docs_dir}{Local path to documentation folder}

\item{year}{Year of ACS data, as number}

\item{span}{1 or 5 year for ACS data, as a number}
}
\description{
For a given year/span (eg. 2018 1yr) gets Seq/Table number data and geogrphy
IDs data and saves them to the provided local directory for documentation
files.
}
