% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fips_abb_name_table}
\alias{fips_abb_name_table}
\title{US State (plus DC and PR) Abbreviations and FIPS codes}
\format{A data frame with 53 rows and 3 variables:
* abb: Official USPS Code
* fips: Fips State Numeric Code
* name: State Name (format used for FTP download files)}
\source{
\url{https://www.census.gov/geo/reference/ansi_statetables.html}
}
\usage{
fips_abb_name_table
}
\description{
A dataset containing state fips coed and USPS abbreviations Used internally
for downloading files, but also useful to create cevtor of state
abbreviations that includes DC and PR.
}
\keyword{datasets}
