% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_acs_vars.R
\name{parse_acs_vars}
\alias{parse_acs_vars}
\title{Parse and R script for ACS Variables Used}
\usage{
parse_acs_vars(file)
}
\arguments{
\item{file}{Path to one or more R scripts used to create your derived variables from
ACS-SF data.
Should include ACS variables using the format \code{b01001_e001}
for estimates, \code{b01001_m001} for margins, and \code{b01001_se001} for standard
errors. Use of \code{\link[=acs_sum]{acs_sum()}} to specify variables will also be correctly
parsed.}
}
\description{
This function parses an R script to extract all the ACS variables that are
used within it. These can then be passed to the \code{keep_vars} argument of
\code{\link[=acs_transform]{acs_transform()}}. This helps limit the data that is extracted from raw files
and processed to only what's needed to create your derived variables.
}
