% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sum_level_info}
\alias{sum_level_info}
\title{Summary levels and ACS surveys}
\format{A data frame with 8 rows and 4 variables:
\itemize{
\item geo_type: Geography type
\item sum_level: Summary level code
\item spans: ACS survey spans avaiable (1-year, 5-year)
\item years: Years of ACS surveys available (for 5-year surveys last year of span)
}}
\source{
\url{https://www2.census.gov/acs2011_1yr/summaryfile/ACS_2011_SF_Tech_Doc.pdf}
}
\usage{
sum_level_info
}
\description{
A dataset containing the summary levels supported by \code{\link[=acs_transform]{acs_transform()}} and the ACS surveys for which they are available.
}
\keyword{datasets}
