% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{swap_geo_type}
\alias{swap_geo_type}
\title{Return desired geography type variable from sum level code or name}
\usage{
swap_geo_type(geo, span, type = c("sum_level", "geo_type"))
}
\arguments{
\item{geo}{\[`character(1)`]: census sum level code or geography type name
(eg. `"140"` or `"tract"`). For supported values see [`sum_level_info`]}

\item{type}{\[`character(1)`]: type of geogrphy ID to return ("name", "abb",
or "fips")}
}
\value{
Character(1) of either geography name, abbreviation, or FIPS code.
}
\description{
Return desired geography type variable from sum level code or name
}
