% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{swap_geo_id}
\alias{swap_geo_id}
\title{Return desired geography variable from FIPS or abbreviation}
\usage{
swap_geo_id(geo, id = c("name", "abb", "fips"))
}
\arguments{
\item{geo}{character or numberic, state FIPS code or state (or US) abreviation}

\item{id}{character, type of geogrphy ID to return ("name", "abb", or "fips")}
}
\description{
Return desired geography variable from FIPS or abbreviation
}
\keyword{internal}
