% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{swap_geo_id}
\alias{swap_geo_id}
\title{Return desired geography variable from FIPS or abbreviation}
\usage{
swap_geo_id(geo, span, type = c("name", "abb", "fips"))
}
\arguments{
\item{geo}{[\code{character(1)}]: The 2-letter abbreviation for the state for which data will be
downloaded. For geogrpahies that do not nest within states, use \code{"US"}.}

\item{span}{[\code{integer(1)}]: The span of years for ACS estimates. ACS 1-year, and
5-year surveys are supported.}

\item{type}{[\code{character(1)}]: type of geogrphy ID to return ("name", "abb", or "fips")}
}
\value{
Character(1) of either geography name, abbreviation, or FIPS code.
}
\description{
Return desired geography variable from FIPS or abbreviation
}
