% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_args}
\alias{validate_args}
\title{Validate function arguments}
\usage{
validate_args(endyear, span, overwrite = NULL)
}
\arguments{
\item{endyear}{\[`integer(1)`]: The endyear of the ACS sample. 2005 through 2016 are
available.}

\item{span}{\[`integer(1)`]: The span of years for ACS estimates. ACS 1-year, and
5-year surveys are supported.}

\item{overwrite}{\[`logical(1)`]:  Whether existing versions of these files be overwriten.
Defaults to `FALSE`.}
}
\value{
None
}
\description{
Validate function arguments
}
