% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_args}
\alias{validate_args}
\title{Validate function arguments}
\usage{
validate_args(year, span, overwrite = NULL, sum_levels = NULL)
}
\arguments{
\item{year}{[\code{integer(1)}]: The year of the ACS sample. 2005 through 2018
are available.}

\item{span}{[\code{integer(1)}]: The span of years for ACS estimates. ACS 1-year,
and 5-year surveys are supported.}

\item{overwrite}{[\code{logical(1)}]: Whether existing versions of these files be
overwriten. Defaults to \code{FALSE}.}

\item{sum_levels}{[\code{character}]: Codes for the desiered Summary Levels (eg.
"140" for Tracts). See the built-in dataset \code{sum_level_info} for all
available codes.}
}
\value{
None
}
\description{
Validate function arguments
}
